refactor: move getBundleWeight to utils
